+++ src/createFormState.ts
@@ -1,7 +1,7 @@
 import { createEvent, createStore } from 'effector';
 import { formSubscriptionItems } from 'final-form';
 
-import { isNil, normalizeSubscriptions, pick, notEquals } from './utils';
+import { normalizeSubscriptions, pick, notEquals, normalizeState } from './utils';
 
 import type {
   FormApi as FFFormApi,
@@ -37,16 +37,14 @@ const createFormState = <FormValues, T extends FormSubscription>(config: {
     setValidationPaused: createEvent<boolean>(),
   };
 
-  const $state = createStore<State>(initialState, {
+  const $state = createStore<State>(normalizeState(initialState, subscribeOn), {
     updateFilter: notEquals,
   })
     .on(formStateApi.update, (s, p) => Object.assign({}, s, p))
     .on(formStateApi.setValidationPaused, (s, isValidationPaused) => Object.assign({}, s, { isValidationPaused }));
 
   finalForm.subscribe((x) => {
-    const normalizedState = subscribeOn.reduce((acc, sub) => (isNil(x[sub]) ? { ...acc, [sub]: null } : acc), x);
-
-    formStateApi.update(normalizedState as unknown as Omit<State, 'isValidationPaused'>);
+    formStateApi.update(normalizeState(x, subscribeOn) as unknown as Omit<State, 'isValidationPaused'>);
     // @ts-expect-error
   }, normalizeSubscriptions(formSubscriptionItems, subscribeOn));