+++ src/utils/isNil.ts
@@ -1,0 +1,5 @@
+import { Nil } from '../types';
+
+const isNil = <T>(x: T | Nil): x is Nil => x === null || x === undefined;
+
+export { isNil };